requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
argon2-cffi==23.1.0
bcrypt==4.0.1
bleach==6.1.0 